+++ packages/stream_video_push_notification/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.10.4
+* Sync version with `stream_video_flutter` 0.10.4
+
 ## 0.10.3
 * Sync version with `stream_video_flutter` 0.10.3